Ajustando retornos api NoContent

+++ ProEventos/Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -27,7 +27,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if(eventos == null) return NotFound("Nenhum evento encontrado");
+                if(eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -44,7 +44,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if(evento == null) return NotFound("Evento por id não encontrado");
+                if(evento == null) return NoContent();
                 
                 return Ok(evento);
             }
@@ -61,7 +61,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if(eventos == null) return NotFound("Nenhum evento por tema encontrado");
+                if(eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -78,7 +78,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if(evento == null) return BadRequest("Erro ao tentar adicionar evento");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -95,7 +95,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if(evento == null) return BadRequest("Erro ao tentar atualizar evento");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -111,8 +111,11 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, false);
+                if(evento == null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ? 
-                    Ok("Deletado") : BadRequest("Evento não deletado");
+                    Ok("Deletado") : throw new Exception("Ocorreu um erro ao tentar deletar o evento!");
             }
             catch (Exception ex)
             {
